[Hotfix-21-FCM-초기화-후-재사용] FirebaseApp 초기화 후 유지.
